+++ app/routes/import-message.js
@@ -9,7 +9,8 @@ const outboundMessageConfig = require('../../config/lib/middleware/import-messag
 
 // Middleware
 const paramsMiddleware = require('../../lib/middleware/import-message/params');
-const broadcastMiddleware = require('../../lib/middleware/import-message/broadcast');
+const getBroadcastMiddleware = require('../../lib/middleware/import-message/broadcast-get');
+const parseBroadcastMiddleware = require('../../lib/middleware/import-message/parse-broadcast');
 const getConvoMiddleware = require('../../lib/middleware/conversation-get');
 const createConvoMiddleware = require('../../lib/middleware/conversation-create');
 const updateConvoMiddleware = require('../../lib/middleware/import-message/conversation-update');
@@ -17,7 +18,8 @@ const loadOutboundMessageMiddleware = require('../../lib/middleware/message-outb
 const createOutboundMessageMiddleware = require('../../lib/middleware/message-outbound-create');
 
 router.use(paramsMiddleware());
-router.use(broadcastMiddleware());
+router.use(getBroadcastMiddleware());
+router.use(parseBroadcastMiddleware());
 
 // Load or create conversation
 router.use(getConvoMiddleware());
