--- lib/middleware/import-message/broadcast.js
@@ -1,92 +1,0 @@
-'use strict';
-
-const Promise = require('bluebird');
-const Cacheman = require('cacheman');
-const logger = require('../../logger');
-
-const config = require('../../../config/lib/middleware/import-message/broadcast');
-
-const cache = new Cacheman('broadcasts', { ttl: config.cacheTtl });
-
-const helpers = require('../../helpers');
-const contentful = require('../../contentful');
-const UnprocessibleEntityError = require('../../../app/exceptions/UnprocessibleEntityError');
-
-// TODO: Move inside lib/contentful
-function getCampaignId(broadcastObject) {
-  const campaign = broadcastObject.fields.campaign;
-  if (!campaign.fields) {
-    return null;
-  }
-  return campaign.fields.campaignId;
-}
-
-// TODO: Move inside lib/contentful
-function getTopic(broadcastObject) {
-  return broadcastObject.fields.topic;
-}
-
-// TODO: Move inside lib/contentful
-function getMessage(broadcastObject) {
-  return broadcastObject.fields.message;
-}
-
-// TODO: Move inside lib/contentful
-function parsePropertiesFromBroadcast(req, res, broadcast) {
-  let promise;
-  if (broadcast) {
-    promise = Promise.resolve(broadcast);
-  } else {
-    promise = new Promise((resolve, reject) => {
-      contentful.fetchBroadcast(req.broadcastId)
-        .then((broadcastObj) => {
-          cache.set(req.broadcastId, broadcastObj)
-            .then((cachedBroadcast) => {
-              resolve(cachedBroadcast);
-            })
-            .catch(error => reject(error));
-        })
-        .catch(err => reject(err));
-    });
-  }
-
-  return promise
-    .then((cachedBroadcast) => {
-      req.topic = getTopic(cachedBroadcast);
-      req.campaignId = getCampaignId(cachedBroadcast);
-      req.outboundMessageText = getMessage(cachedBroadcast);
-      return cachedBroadcast;
-    });
-}
-
-module.exports = function getBroadcast() {
-  return (req, res, next) => {
-    const asyncProperties = [];
-    const broadcastId = req.broadcastId;
-
-    // We only import messages sent in broadcasts. For this reason, broadcastId must be
-    // a required field. If in the future we want to allow generic imports we can relax
-    // this requirement here and call next()
-    if (!broadcastId) {
-      const error = new UnprocessibleEntityError('broadcastId is a required property.');
-      return helpers.sendErrorResponse(res, error);
-    }
-    helpers.analytics.addParameters({ broadcastId });
-
-    return cache.get(broadcastId)
-      .then((broadcast) => {
-        if (broadcast) {
-          logger.debug('Broadcasts cache hit', { broadcastId: req.broadcastId }, req);
-          asyncProperties.push(parsePropertiesFromBroadcast(req, res, broadcast));
-        } else {
-          logger.debug('Broadcasts cache miss', { broadcastId: req.broadcastId }, req);
-          asyncProperties.push(parsePropertiesFromBroadcast(req, res));
-        }
-
-        // Process async parsers
-        return Promise.all(asyncProperties)
-          .then(() => next())
-          .catch(err => helpers.sendErrorResponse(res, err));
-      });
-  };
-};
